requirements.txt
playwright==1.48.2
pytest==7.4.4
pytest-playwright==0.5.1
pytest-asyncio==0.23.3
python-dotenv==1.0.0